ACHILLES_HEEL_results.record_count -> BIGINT
Change ACHILLES_HEEL_results.record_count type from INT to  BIGINT

+++ inst/sql/sql_server/AchillesHeel_v5.sql
@@ -52,7 +52,7 @@ CREATE TABLE @results_database_schema.ACHILLES_HEEL_results (
   analysis_id INT,
 	ACHILLES_HEEL_warning VARCHAR(255),
 	rule_id INT,
-	record_count INT
+	record_count BIGINT
 );
 
 --ruleid 1 check for non-zero counts from checks of improper data (invalid ids, out-of-bound data, inconsistent dates)
